保存light测试进度

+++ LumiRender/src/tests/test_reflection.cpp
@@ -78,8 +78,8 @@ int main() {
 //        std::cout << typeid(T).name() << std::endl;
 //    });
 
-    uint32_t i;
-    printf("%u\n", i);
+    float i;
+    printf("%f\n", i);
 
     for_each_all_registered_member<D<int>>([&](auto offset, auto name, auto ptr) {
         cout << typeid(ptr).name() << "  " << name << endl;
